requirements.txt
Django==3.2
django-cors-headers==3.13.0
django-crispy-forms==1.9.2
django-environ==0.9.0
django-filter==22.1
django-heroku==0.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
nltk==3.7
pymongo==3.12.3
beautifulsoup4==4.10.0
requests==2.27.1
gunicorn==20.0.4
pytz==2020.1
asgiref==3.3.2
coverage==5.5